+++ CHANGES.rst
@@ -22,6 +22,7 @@ Changelog
   base class.
 * Add a new ``update`` CLI sub-command.
 * Add a new ``sync`` CLI sub-command.
+* Rename managers' ``active`` property to ``available``.
 
 
 `1.8.0 (2016-08-22) <https://github.com/kdeldycke/meta-package-manager/compare/v1.7.0...v1.8.0>`_